resource/aws_redshift_cluster: Send correct values when resizing
When resizing a Redshift cluster, the AWS API is expecting the Cluster
Type, Node Type, and Number of Nodes regardless of which value changed.
If not passed it will error out with
`[WARN] Error modifying Redshift Cluster (navistone-public-cluster):
InvalidParameterCombination: Number of nodes for cluster type multi-node
must be greater than or equal to 2`

Now we send all values if any one of them change.

Signed-off-by: Ken Herner <kherner@navistone.com>